requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
gunicorn>=21.0
whitenoise>=6.0
psycopg2-binary>=2.9
dj-database-url>=2.0
django-allauth>=65.0
requests>=2.31
PyJWT>=2.8
cryptography>=41.0
djangorestframework>=3.15
drf-spectacular>=0.27
django-filter>=24.0
